// src/types/types.ts
export type PaymentResponse = {
  id: number;
  caseBook: string;
  responseStatus: string;
  transactionDate: string;
  transactionReferenceNumber: string;
  currentAmount: string;
  exchangeRate: string;
  paymentChannelReference: string;
  swiftMsgCategorySubType: string;
  swiftMsgType: string;
  paymentDate: string;
  stpFlag: string;
  creditPartyAcct: string;
  creditPartyName: string;
  creditPartyAddr1: string;
  creditPartyAddr2: string;
  creditPartyAddr3: string;
  creditAmount: string;
  creditValueDate: string;
  beneficiaryName: string;
  beneficiaryAddr1: string;
  beneficiaryAddr2: string;
  beneficiaryAddr3: string;
  remittanceInfo2: string;
  remittanceInfo3: string;
  remittanceInfo4: string;
  debitPartyAcct: string;
  debitPartyName: string;
  debitPartyAddr1: string;
  debitPartyAddr2: string;
  debitPartyAddr3: string;
  debitAmount: string;
  debitValueDate: string;
  sendingBankIdType: string;
  sendingBankId: string;
  sendingBankName: string;
  sendingBankAddr1: string;
  sendingBankAddr2: string;
  sendingBankAddr3: string;
  sendingBankTranRef: string;
  orderingBankIdType: string;
  orderingBankId: string;
  orderingBankName: string;
  orderingBankAddr1: string;
  orderingBankAddr2: string;
  orderingBankAddr3: string;
  createdBy: string;
  createdDateTime: string;
  transactionStatus: string;
  swiftInMsgIdentifier: string;
  swiftInMsgType: string;
  swiftInField20: string;
  swiftInField21: string;
  swiftInCurrAmount: string;
  swiftInValueDate: string;
  swiftInCbaBic: string;
  swiftInExternalPartyBic: string;
  swiftInStatus: string;
  swiftInSwiftInMsgIdentifier: string;
  swiftInMsgSequence: string;
  swiftInDateTime: string;
  swiftInSource: string;
  swiftInMsgResponse: string;
  swiftInSessionNo: string;
  swiftInSwiftMsg: string;
  swiftOutMsgIdentifier: string;
  swiftOutMsgType: string;
  swiftOutField20: string;
  swiftOutField21: string;
  swiftOutCurrAmount: string;
  swiftOutValueDate: string;
  swiftOutRemitter: string;
  swiftOutBnp: string;
  swiftOutCbaBic: string;
  swiftOutExternalPartyBic: string;
  swiftOutStatus: string;
  swiftOutMsgId: string;
  swiftOutMsgSequence: string;
  swiftOutDateTime: string;
  swiftOutSource: string;
  swiftOutMsgResponse: string;
  swiftOutSessionNo: string;
  swiftOutCorrespondentCtry: string;
  swiftOutMsgUserRef: string;
  swiftOutCreatedBy: string;
  swiftOutCreatedDateTime: string;
  swiftMsg: string;
  swiftCreatedBy: string;
  swiftCreatedDateTime: string;
  swiftOut2: string;
  swiftInGp2_199: string;
  swiftOutGgp2_199: string;
};